delete warning

+++ src/hammerfest/level/ennemies/Ennemie.cpp
@@ -6,7 +6,7 @@ static const int accelerate[tableSize] = { 2, 3, 5, 10 };
 
 Ennemie::Ennemie(int id, int x, int y, int type, Level * level) :
 	Position((x * 20) + 10, (y * 20) + 20), Drawable(), HitBox(), IdElement(id) {
-	std::srand(std::time(NULL));
+	std::srand((int)std::time(NULL));
 	this->type = type;
 	isAngry = false;
 	changeState(walk);
